refactor(ui): group diff worker internals
