+++ src/highlightWorker.ts
@@ -1,2 +1,2 @@
 /** Exposes the UI-local highlighting worker as a compiled Bun entrypoint. */
-import "./ui/diff/highlightWorker";
+import "./ui/diff/worker/highlightWorker";
