+++ src/ui/diff/worker/highlightHast.ts
