+++ src/ui/diff/worker/index.ts
@@ -1,0 +1,18 @@
+/** Exposes the terminal diff worker subsystem without requiring callers to know its internals. */
+export { createHighlightWorker } from "../../../highlightWorkerClient";
+export {
+  disposeHighlightWorker,
+  highlightDiffInWorker,
+  registerHighlightWorker,
+  type WorkerHighlightedDiffCode,
+} from "./highlightWorkerClient";
+export {
+  compactHighlightRunsForLine,
+  compactHighlightTransferList,
+  compactHighlightedDiffByteLength,
+  encodeCompactHighlightedDiff,
+  validateCompactHighlightedDiff,
+  type CompactHighlightedDiff,
+  type CompactHighlightRun,
+} from "./highlightCompact";
+export { collectHastHighlightRuns, type HastNode } from "./highlightHast";
